refactor: move more my-units files
