+++ apps/admin-ui/src/App.tsx
@@ -9,7 +9,7 @@ import { PUBLIC_URL } from "./common/const";
 import { GlobalContext } from "./context/GlobalContexts";
 import { prefixes } from "./common/urls";
 import AuthorizationChecker from "./common/AuthorizationChecker";
-import MyUnitsRouter from "./spa/my-units/MyUnitsRouter";
+import MyUnitsRouter from "./spa/my-units/router";
 import ReservationsRouter from "./component/reservations/ReservationRouter";
 import NotificationsRouter from "./component/notifications/router";
 import Error404 from "./common/Error404";
