+++ apps/admin-ui/src/spa/my-units/[id]/eventStyleGetter.ts
