--- apps/admin-ui/src/spa/my-units/create-reservation/queries.tsx
@@ -1,11 +1,0 @@
-import { gql } from "@apollo/client";
-
-export const CREATE_STAFF_RESERVATION = gql`
-  mutation CreateStaffReservation(
-    $input: ReservationStaffCreateMutationInput!
-  ) {
-    createStaffReservation(input: $input) {
-      pk
-    }
-  }
-`;
