+++ apps/admin-ui/src/spa/my-units/router.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
 import { RecurringReservation } from "./recurring/RecurringReservation";
-import { MyUnits } from "./MyUnits";
-import { MyUnitView } from "./MyUnitView";
+import { MyUnits } from "./index";
+import { MyUnitView } from "./[id]/index";
 import { RecurringReservationDone } from "./recurring/RecurringReservationDone";
 
 function MyUnitsRouter(): JSX.Element {
@@ -10,9 +10,14 @@ function MyUnitsRouter(): JSX.Element {
     <Routes>
       <Route index element={<MyUnits />} />
       <Route path=":unitId" element={<MyUnitView />} />
+      <Route path=":unitId/recurring" element={<RecurringReservation />} />
       <Route
         path=":unitId/recurring-reservation"
         element={<RecurringReservation />}
+      />
+      <Route
+        path=":unitId/recurring/completed"
+        element={<RecurringReservationDone />}
       />
       <Route
         path=":unitId/recurring-reservation/completed"
